# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Project to study golang",
        "title": "Simple Ecommerce Go",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bruno",
            "url": "https://github.com/brunoeduardodev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
        "/products": {
            "get": {
                "description": "List all products",
                "produces": [
                    "application/json"
                ],
                "summary": "List all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/views.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product",
                "produces": [
                    "application/json"
                ],
                "summary": "Create Product",
                "parameters": [
                    {
                        "description": "New product input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sign-in": {
            "post": {
                "description": "Does a sign in",
                "produces": [
                    "application/json"
                ],
                "summary": "Does a sign in",
                "parameters": [
                    {
                        "description": "Sign In input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SignInInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/sign-up": {
            "post": {
                "description": "Does a sign up",
                "produces": [
                    "application/json"
                ],
                "summary": "Does a sign up",
                "parameters": [
                    {
                        "description": "Sign up input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repositories.CreateUserInput"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "controllers.SignInInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "repositories.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "services.CreateProductInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "views.Product": {
            "description": "Information of a product",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "string": {
                    "description": "Created at here",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}